+++ src/main/java/agents/firm/production/control/facades/DiscreteMatcherPlantControl.java
@@ -63,7 +63,7 @@ public class DiscreteMatcherPlantControl  implements PlantControl
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(GoodType type) {
+    public int maxPrice(GoodType type) {
         return control.maxPrice(type);
     }
 
@@ -85,7 +85,7 @@ public class DiscreteMatcherPlantControl  implements PlantControl
      * @param newWage the new wage
      */
     @Override
-    public void setCurrentWage(long newWage) {
+    public void setCurrentWage(int newWage) {
         control.setCurrentWage(newWage);
     }
 
@@ -147,7 +147,7 @@ public class DiscreteMatcherPlantControl  implements PlantControl
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(Good good) {
+    public int maxPrice(Good good) {
         return control.maxPrice(good);
     }
