+++ src/main/java/agents/firm/production/control/facades/DumbClimberControl.java
@@ -80,7 +80,7 @@ public class DumbClimberControl implements PlantControl
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(GoodType type) {
+    public int maxPrice(GoodType type) {
         return control.maxPrice(type);
     }
 
@@ -89,7 +89,7 @@ public class DumbClimberControl implements PlantControl
      * @param newWage the new wage
      */
     @Override
-    public void setCurrentWage(long newWage) {
+    public void setCurrentWage(int newWage) {
         control.setCurrentWage(newWage);
     }
 
@@ -175,7 +175,7 @@ public class DumbClimberControl implements PlantControl
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(Good good) {
+    public int maxPrice(Good good) {
         return control.maxPrice(good);
     }
