+++ src/main/java/agents/firm/production/control/facades/FixWagesPlantControl.java
@@ -62,7 +62,7 @@ public class FixWagesPlantControl implements PlantControl {
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(GoodType type) {
+    public int maxPrice(GoodType type) {
         return control.maxPrice(type);
     }
 
@@ -84,7 +84,7 @@ public class FixWagesPlantControl implements PlantControl {
      * @param newWage the new wage
      */
     @Override
-    public void setCurrentWage(long newWage) {
+    public void setCurrentWage(int newWage) {
         control.setCurrentWage(newWage);
     }
 
@@ -146,7 +146,7 @@ public class FixWagesPlantControl implements PlantControl {
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(Good good) {
+    public int maxPrice(Good good) {
         return control.maxPrice(good);
     }
