+++ src/main/java/agents/firm/production/control/facades/LookAheadAnnealingControl.java
@@ -149,7 +149,7 @@ public class LookAheadAnnealingControl implements PlantControl
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice( GoodType type) {
+    public int maxPrice(GoodType type) {
         return control.maxPrice(type);
     }
 
@@ -160,7 +160,7 @@ public class LookAheadAnnealingControl implements PlantControl
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice( Good good) {
+    public int maxPrice(Good good) {
         return control.maxPrice(good);
     }
 
@@ -177,7 +177,7 @@ public class LookAheadAnnealingControl implements PlantControl
      * @param newWage the new wage
      */
     @Override
-    public void setCurrentWage(long newWage) {
+    public void setCurrentWage(int newWage) {
         control.setCurrentWage(newWage);
     }
