+++ src/main/java/agents/firm/production/control/facades/MarginalPlantControl.java
@@ -162,7 +162,7 @@ public class MarginalPlantControl implements PlantControl, PlantListener {
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice( GoodType type) {
+    public int maxPrice(GoodType type) {
         return control.maxPrice(type);
     }
 
@@ -173,7 +173,7 @@ public class MarginalPlantControl implements PlantControl, PlantListener {
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice( Good good) {
+    public int maxPrice(Good good) {
         return control.maxPrice(good);
     }
 
@@ -190,7 +190,7 @@ public class MarginalPlantControl implements PlantControl, PlantListener {
      * @param newWage the new wage
      */
     @Override
-    public void setCurrentWage(long newWage) {
+    public void setCurrentWage(int newWage) {
         control.setCurrentWage(newWage);
     }
