+++ src/main/java/agents/firm/production/control/facades/MarginalPlantControlWithPIDUnit.java
@@ -174,7 +174,7 @@ public class MarginalPlantControlWithPIDUnit implements PlantControl, PlantListe
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice( GoodType type) {
+    public int maxPrice(GoodType type) {
         return control.maxPrice(type);
     }
 
@@ -185,7 +185,7 @@ public class MarginalPlantControlWithPIDUnit implements PlantControl, PlantListe
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice( Good good) {
+    public int maxPrice(Good good) {
         return control.maxPrice(good);
     }
 
@@ -202,7 +202,7 @@ public class MarginalPlantControlWithPIDUnit implements PlantControl, PlantListe
      * @param newWage the new wage
      */
     @Override
-    public void setCurrentWage(long newWage) {
+    public void setCurrentWage(int newWage) {
         control.setCurrentWage(newWage);
     }
