+++ src/main/java/agents/firm/production/technology/DRSExponentialMachinery.java
@@ -33,12 +33,12 @@ public class DRSExponentialMachinery extends ExponentialMachinery {
 
         private float alpha = 0.1f;
 
-    public DRSExponentialMachinery( GoodType type,  EconomicAgent producer, long costOfProduction,  Plant plant, float outputMultiplier, float alpha) {
+    public DRSExponentialMachinery( GoodType type,  EconomicAgent producer, int costOfProduction,  Plant plant, float outputMultiplier, float alpha) {
         super(type, producer, costOfProduction, plant, outputMultiplier);
         this.alpha = alpha;
     }
 
-    public DRSExponentialMachinery( GoodType type,  EconomicAgent producer, long costOfProduction,  Plant plant, float alpha) {
+    public DRSExponentialMachinery( GoodType type,  EconomicAgent producer, int costOfProduction,  Plant plant, float alpha) {
         super(type, producer, costOfProduction, plant);
         this.alpha = alpha;
     }
