+++ src/main/java/agents/firm/purchases/pid/PurchasesFixedPID.java
@@ -165,7 +165,7 @@ public class PurchasesFixedPID extends FixedInventoryControl implements BidPrici
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(GoodType type) {
+    public int maxPrice(GoodType type) {
         return Math.round(controller.getCurrentMV());
     }
 
@@ -176,7 +176,7 @@ public class PurchasesFixedPID extends FixedInventoryControl implements BidPrici
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(Good good) {
+    public int maxPrice(Good good) {
         return maxPrice(good.getType());
     }
