+++ src/main/java/agents/firm/purchases/pid/PurchasesSimplePID.java
@@ -113,7 +113,7 @@ public class PurchasesSimplePID extends SimpleInventoryControl implements BidPri
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(GoodType type) {
+    public int maxPrice(GoodType type) {
         return Math.round(controller.getCurrentMV());
     }
 
@@ -124,7 +124,7 @@ public class PurchasesSimplePID extends SimpleInventoryControl implements BidPri
      * @return the maximum price I am willing to pay for this good
      */
     @Override
-    public long maxPrice(Good good) {
+    public int maxPrice(Good good) {
         return maxPrice(good.getType());
     }
